More auth and user tests

+++ internal/http/api/controllers/v1/users/users.go
@@ -516,6 +516,10 @@ func DELETEUser(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid user ID"})
 		return
 	}
+	if !models.UserIDExists(db, uint(idUint64)) {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "User does not exist"})
+		return
+	}
 	models.DeleteUser(db, uint(idUint64))
 	if db.Error != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": db.Error.Error()})
